refactor(logic): split work-tile order suggestion helpers (Refs #2288)
Replace the monolithic order_suggestion_build_research.dart with focused
order_suggestion_work_tile_keys.dart (valid work-tile APIs and visibility
sorting) and order_suggestion_work_tile_prefilter.dart (raw candidate tiles
and work-target prefilter map). suggestBuildOrders / suggestResearchOrders
stay in their existing dedicated files.

Update barrel exports, order_suggestion_work library imports, and
logic_all_provinces_sanctions.yaml paths for the moved allProvinces call sites.
