+++ packages/colonizethis_logic/lib/src/orders/order_suggestion_work.dart
@@ -8,7 +8,8 @@ import '../world/player_view.dart';
 import '../world/province_lookup.dart';
 import '../world/unit_lookup.dart';
 import 'bundled_civilian_work_order.dart';
-import 'order_suggestion_build_research.dart';
+import 'order_suggestion_work_tile_keys.dart';
+import 'order_suggestion_work_tile_prefilter.dart';
 import 'order_suggestion_context.dart';
 import 'order_visibility.dart';
 import 'partial_province_reveal.dart';
