+++ tool/logic_all_provinces_sanctions.yaml
@@ -45,14 +45,14 @@ sanctions:
     spec: SPEC/program/logic-dual-region-province-access.md
     issue: https://github.com/waigore/colonizethisv3/issues/2278
 
-  - path: packages/colonizethis_logic/lib/src/orders/order_suggestion_build_research.dart
-    line: 255
+  - path: packages/colonizethis_logic/lib/src/orders/order_suggestion_work_tile_prefilter.dart
+    line: 52
     rationale: Build suggestions scan province context across regions.
     spec: SPEC/program/logic-dual-region-province-access.md
     issue: https://github.com/waigore/colonizethisv3/issues/2278
 
-  - path: packages/colonizethis_logic/lib/src/orders/order_suggestion_build_research.dart
-    line: 328
+  - path: packages/colonizethis_logic/lib/src/orders/order_suggestion_work_tile_prefilter.dart
+    line: 109
     rationale: Research/build suggestion enumeration over all provinces.
     spec: SPEC/program/logic-dual-region-province-access.md
     issue: https://github.com/waigore/colonizethisv3/issues/2278
